+++ packages/@borealisgroup/theme/src/index.js
@@ -1,7 +1,0 @@
-import borealis from './brands/borealis.json';
-import ecoplast from './brands/ecoplast.json';
-import lat from './brands/lat.json';
-import mtmPlastics from './brands/mtmPlastics.json';
-import rosier from './brands/rosier.json';
-
-export { borealis, lat, ecoplast, mtmPlastics, rosier };