+++ packages/@borealisgroup/theme/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.1](https://github.com/borealisgroup/borealis/tree/master/packages/@borealisgroup/theme/compare/@borealisgroup/theme@1.9.0...@borealisgroup/theme@1.9.1) (2019-12-10)
+
+
+### Bug Fixes
+
+* files dist ([ee44300](https://github.com/borealisgroup/borealis/tree/master/packages/@borealisgroup/theme/commit/ee44300cec0fa16a394964261baa308f1c863331))
+
+
+
+
+
+# [1.9.0](https://github.com/borealisgroup/borealis/tree/master/packages/@borealisgroup/theme/compare/@borealisgroup/theme@1.8.1...@borealisgroup/theme@1.9.0) (2019-12-10)
+
+
+### Features
+
+* types ([b35eacf](https://github.com/borealisgroup/borealis/tree/master/packages/@borealisgroup/theme/commit/b35eacf4bdda3bb77bfbe35f66661c5e5ad3e2e8))
+
+
+
+
+
 ## [1.8.1](https://github.com/borealisgroup/borealis/tree/master/packages/@borealisgroup/theme/compare/@borealisgroup/theme@1.8.0...@borealisgroup/theme@1.8.1) (2019-12-02)
 
 **Note:** Version bump only for package @borealisgroup/theme